constants rearrange
