+++ internal/botclass.py
@@ -2,6 +2,7 @@ import asyncio
 import logging
 from pathlib import Path
 
+# noinspection PyPackageRequirements
 import discord
 from discord.ext import commands
 from pretty_help import PrettyHelp
