+++ internal/constants_bot_prod.py
@@ -1,0 +1,36 @@
+"""Constants for bot use."""
+
+# Customize the following
+
+ROLE_WHITELIST = [
+    699145313520320542,  # LIVE - Admin
+    725198459627634689,  # LIVE - Mod
+    758399333942558800,  # LIVE - Tourny Org
+    808426677562114122,  # LIVE - Record Org
+]
+MAP_CHANNEL_ID = 795399258966458459
+RECORD_CHANNEL_ID = 693673770086301737
+MAP_SUBMIT_CHANNEL_ID = 689621322862559413
+HIDDEN_VERIFICATION_CHANNEL = 813768098191769640
+
+GUILD_ID = 689587520496730129
+
+BOT_ID = 801483463642841150
+
+# The amount of servers required for a bot to be eligible for verification.
+
+# The % of bot accounts in a server in order for the server to be considered low quality
+
+# Message sent at the bottom of every reaction confirmaation
+
+# Verification
+
+# Prettytable page size
+
+# creator / desc max_lengths
+
+# limit for latest maps command
+
+# Map names | CUSTOMIZE ACCEPTABLE NAMES
+
+# combined map names
