+++ internal/constants_bot_test.py
@@ -1,0 +1,16 @@
+"""Constants for use with 'test' arg."""
+
+# Customize the following
+
+ROLE_WHITELIST = [
+    195542852518805504,  # TEST - Mod
+    801645674617634886,  # TEST - Tester
+]
+MAP_CHANNEL_ID = 802362144506511400
+RECORD_CHANNEL_ID = 801496775390527548
+MAP_SUBMIT_CHANNEL_ID = 802624308726726707
+HIDDEN_VERIFICATION_CHANNEL = 811467249100652586
+
+GUILD_ID = 195387617972322306
+
+BOT_ID = 808340225460928552
